Refactor : Home page
