+++ src/views/List/hooks/useGetUserData.tsx
@@ -1,16 +1,15 @@
 import { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import userDataService from '../../../service/userData/userData';
 import { selectIsLoggedIn } from '../../../features/user/selectors';
-import { initStockList } from '../../../features/stockList/stockListSlice';
-import { initCheckedItems } from '../../../features/checkedItems/checkedItemsSlice';
-import { initGroups } from '../../../features/groups/groupsSlice';
+import useInitUserData from '../../../features/useInitUserData';
 
 const useGetUserData = () => {
-  const dispatch = useDispatch();
   const isLoggingIn = useSelector(selectIsLoggedIn);
   const [loading, setLoading] = useState(false);
+  const initUserData = useInitUserData();
+
   const getUserData = async () => {
     setLoading(true);
 
@@ -25,9 +24,8 @@ const useGetUserData = () => {
       return;
     }
 
-    dispatch(initStockList(result.stocks));
-    dispatch(initCheckedItems(result.checkedItems));
-    dispatch(initGroups(result.groups));
+    initUserData(result);
+
     setLoading(false);
   };
